refactor: extract discord client
