--- src/client.js
@@ -1,7 +1,0 @@
-const { Client, Intents } = require('discord.js')
-
-const newClient = () => new Client({ intents: [Intents.FLAGS.GUILDS] })
-
-module.exports = {
-  newClient,
-}
